+++ src/GRA.Controllers/MissionControl/FlightController.cs
@@ -8,7 +8,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Threading.Tasks;
-using System.Collections.Generic;
 using System.Linq;
 
 namespace GRA.Controllers.MissionControl